add new Identifiable interface
The new emcomapi.Identifiable type is a low-level type (w.r.t. hierarchy of components in the emersyx platform) for all components which need to be uniquely identifiable. This includes resources, receptors, gateways (i.e. types which are simultaneously resources and receptors) and processors. This type is to be used internally by emersyx core.