router api changes
I have made changes to the router API in order to follow the same structure as for the IRC and Telegram gateway APIs. I added a new RouterOptions interface with options relevant to routers, removed the overlapping methods from the emrtrapi.Router interface, created a NewRouterOptions function and changed the signature of the NewRouter function.