distinguish input and output channels for processors
There was no way for processors to push events (as described in the architecture document). To fix this, in the emcomapi.Processor interface, I renamed the GetEventsChannel method to GetInEventsChannel and added an additional GetOutEventsChannel method. These two methods allow router implementations to distinguish between the two channels used to send and receive events.