change router api to allow gateways and processors to be added after creation
In the previous API, a Router object could only be created if there were already Processor objects. Processors objects could only be created if there was a router object. This commit changes the router API such that options (i.e. gateways, processors and routes) can be added after the object is created. Thus the workflow becomes 1) create an empty router, 2) create processors and gateways and 3) configure the router with the created processors and gateways.