add NewTelegramParameters method to TelegramResource interface
The methods of the TelegramResource interface make use of the TelegramParameters type for arguments. The TelegramParameters type is linked to a specific TelegramResource/TelegramGateway implementation. As such, an emcomapi.Processor needs to access the correct TelegramParameters implementation when making use of the gateway. In order to do this, the processor either needs access to information about the plugin file which implements the TelegramGateway, or just ask the TelegramGateway directly to generate a new TelegramParameters instance. The latter is easier to implement, and easier from a logistics point of view.